Windows port: experimental ccl: MSVC narrowing casts and pimpl-completeness fixes
- all_gather_async, all_gather_concat_heads_fused, all_gather_minimal_matmul_async,
  all_reduce_async, all_to_all_async(_generic), all_to_all_dispatch_metadata,
  deepseek_moe_reduce_scatter, llama_all_gather_matmul_async, llama_reduce_scatter(_create_heads),
  llama_reduce_scatter_matmul, minimal_matmul_strided_reduce_scatter_async, moe_compute,
  neighbor_pad_async, neighbor_pad_halo, reduce_scatter_minimal_async(_direct),
  ring_attention_all_gather_async, send_recv_async (recv/send, direct and indirect),
  slice_reshard_async, strided_all_gather_async, strided_reduce_scatter_async:
  add static_cast<uint32_t>(...) around DeviceAddr/size_t/int expressions
  (.buffer()->address(), .address(), .size(), CoreCoord.x/.y, etc.) that MSVC treats
  as narrowing inside std::vector<uint32_t> and other braced-init-lists used for
  compile-time and runtime kernel args.
- ring_attention_all_gather_async: fix two CBDescriptor aggregate-init failures
  (C2440) caused by size_t packet-header-size expressions narrowing into
  uint32_t total_size/page_size fields.
- all_gather_async, all_to_all_dispatch_metadata, llama_reduce_scatter(_create_heads),
  minimal_matmul_strided_reduce_scatter_async, moe/selective_reduce_combine,
  reduce_scatter_minimal_async, ring_attention_all_gather_async, strided_all_gather_async,
  strided_reduce_scatter_async, dit_fused_distributed_groupnorm/rmsnorm,
  all_gather_minimal_matmul_async: include tt_metal/impl/buffers/global_semaphore_impl.hpp
  wherever operation_attributes_t (or a struct it nests) stores a GlobalSemaphore by
  value (in std::optional or std::vector). MSVC's aggregate/attribute reflection
  machinery in device_operation.hpp and tt_stl/reflection.hpp needs the
  GlobalSemaphoreImpl pimpl type to be complete at these points; GCC/Clang do not.

+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_to_all_dispatch_metadata/device/all_to_all_dispatch_metadata_program_factory.cpp
@@ -529,7 +529,7 @@ AllToAllDispatchMetadataDeviceOperation::AllToAllDispatchMetadataSparse::create_
         batch_size,          // 17
         selected_experts_k,  // 18
         experts,             // 19
-        num_shared_experts,  // 20
+        static_cast<uint32_t>(num_shared_experts),  // 20
         tokens_per_device,   // 21
 
         num_links,           // 22
@@ -537,8 +537,8 @@ AllToAllDispatchMetadataDeviceOperation::AllToAllDispatchMetadataSparse::create_
 
         src_mesh_id,            // 24
         (uint32_t)src_chip_id,  // 25
-        mesh_view.num_rows(),   // 26
-        mesh_view.num_cols(),   // 27
+        static_cast<uint32_t>(mesh_view.num_rows()),   // 26
+        static_cast<uint32_t>(mesh_view.num_cols()),   // 27
 
         aligned_input_page_size,     // 28
         aligned_indices_page_size,   // 29
